refactor: get rid of useless record classes
There's no point in having separate FooInit and ConstFoo classes

+++ src/main/java/com/quaxt/mcc/asm/Operand.java
@@ -3,7 +3,7 @@ package com.quaxt.mcc.asm;
 import com.quaxt.mcc.Err;
 
 public sealed interface Operand permits Data, DoubleReg, Imm, Pseudo, Reg, Memory, PseudoMem, Indexed {
-    default Operand plus(int offset) {
+    default Operand plus(long offset) {
         throw new Err("Operation not supported");
     }
 }
